src/containers/analytics-pyspark/requirements.txt
pyspark
delta-spark
pandas
matplotlib
